autogen-example/agents/requirements.txt
fastapi
autogen-core
autogen-ext[openai]
autogen-ext[azure]
autogen-agentchat
confluent_kafka
uvicorn
python-dotenv
pymongo
beautifulsoup4